+++ tests/unit/factories.py
@@ -8,7 +8,12 @@ from unittest.mock import MagicMock
 
 import factory
 
-from src.state import BACKENDS_CONFIG_NAME, LOCATION_CONFIG_NAME, PROTOCOL_CONFIG_NAME
+from src.state import (
+    BACKENDS_CONFIG_NAME,
+    HOSTNAME_CONFIG_NAME,
+    PATH_CONFIG_NAME,
+    PROTOCOL_CONFIG_NAME,
+)
 
 T = typing.TypeVar("T")
 
@@ -41,7 +46,8 @@ class MockCharmFactory(factory.Factory):
     unit = MockUnitFactory
     config = factory.Dict(
         {
-            LOCATION_CONFIG_NAME: "example.com",
+            HOSTNAME_CONFIG_NAME: "example.com",
+            PATH_CONFIG_NAME: "/",
             BACKENDS_CONFIG_NAME: "10.10.1.1, 10.10.2.2",
             PROTOCOL_CONFIG_NAME: "https",
         }